+++ crates/rattler-bin/src/commands/mod.rs
@@ -1,5 +1,8 @@
 pub mod auth;
+pub mod client;
+pub mod completion;
 pub mod create;
+pub mod download;
 pub mod extract;
 pub mod fetch_file;
 pub mod inspect;
@@ -7,4 +10,5 @@ pub mod link;
 pub mod list;
 pub mod menu;
 pub mod search;
+pub mod shell_hook;
 pub mod virtual_packages;