+++ crates/rattler-bin/src/main.rs
@@ -42,13 +42,16 @@ struct Opt {
 #[derive(Debug, clap::Subcommand)]
 enum Command {
     Auth(commands::auth::Opt),
+    Completion(commands::completion::Opt),
     Create(commands::create::Opt),
+    Download(commands::download::Opt),
     FetchFile(commands::fetch_file::Opt),
     Inspect(commands::inspect::Opt),
     Search(commands::search::Opt),
+    ShellHook(commands::shell_hook::Opt),
     VirtualPackages(commands::virtual_packages::Opt),
     InstallMenu(commands::menu::InstallOpt),
-    RemoveMenu(commands::menu::InstallOpt),
+    RemoveMenu(commands::menu::RemoveOpt),
     Extract(commands::extract::Opt),
     Link(commands::link::Opt),
     Upload(Box<rattler_upload::upload::opt::UploadOpts>),
@@ -101,11 +104,14 @@ async fn async_main() -> miette::Result<()> {
     // Dispatch the selected comment
     match opt.command {
         Command::Auth(opts) => commands::auth::auth(opts).await,
+        Command::Completion(opts) => commands::completion::completion(opts),
         Command::Create(opts) => commands::create::create(opts).await,
+        Command::Download(opts) => commands::download::download(opts).await,
         Command::FetchFile(opts) => commands::fetch_file::fetch_file(opts).await,
         Command::Inspect(opts) => commands::inspect::inspect(opts).await,
         Command::Search(opts) => commands::search::search(opts).await,
         Command::List(opts) => commands::list::list(opts).await,
+        Command::ShellHook(opts) => commands::shell_hook::shell_hook(opts).await,
         Command::VirtualPackages(opts) => commands::virtual_packages::virtual_packages(opts),
         Command::InstallMenu(opts) => commands::menu::install_menu(opts).await,
         Command::RemoveMenu(opts) => commands::menu::remove_menu(opts).await,