+++ crates/rattler_solve/benches/sorting_bench.rs
@@ -36,7 +36,6 @@ fn bench_sort(c: &mut Criterion, sparse_repo_data: &SparseRepoData, spec: &str) 
                     None,
                     ChannelPriority::default(),
                     None,
-                    None,
                     rattler_solve::SolveStrategy::Highest,
                     Vec::new(),
                 )