+++ tools/create-resolvo-snapshot/src/main.rs
@@ -76,7 +76,6 @@ async fn main() {
         None,
         ChannelPriority::default(),
         None,
-        None,
         SolveStrategy::default(),
         Vec::new(),
     )